3 interface's realizations were changed

+++ field_logic/Dot.java
@@ -3,19 +3,19 @@ package field_logic;
 import field_logic.intarfaces.iDot;
 
 public class Dot implements iDot {
-    private double xPos;
-    private double yPos;
+    private int xPos;
+    private int yPos;
 
-    Dot(double x, double y) {
+    Dot(int x, int y) {
         xPos = x;
         yPos = y;
     }
 
-    public double getX() {
+    public int getX() {
         return xPos;
     }
 
-    public double getY() {
+    public int getY() {
         return yPos;
     }
 }
